requirements.txt
streamlit
snowflake-connector-python
pandas
python-dotenv
langchain
langchain-groq